Reverse the order used by in_
Currently in_ builds a expression like:

  a OR (b OR (c OR d))

This association causes PostgreSQL's parser to run out of memory for
large lists. For example, calling in_ with 5000 elements will generally
produce an expression that cannot be parsed. The fix is to reassociate
this list:

  ((a OR b) OR c) OR d

PostgreSQL is much happier parsing this, and will successfully parse a
list of 5000 elements.