Allow for schema-qualified table references
When the `tableSchema` is defined, refer to the table using the
schema-qualified table name. By doing this tables in other schemas may
be queried instead of adjusting the defined `search_path` for the
database connections.

Opaleye already supports this and this PR surfaces this functionality
in the definition for `BaseTable`.