Drop columnar arrayAgg (listAgg and nonEmptyAgg are now both tabular)
This is a welcome simplification and eases some of the nonsense we were accumulating in `ExprType` and `ResultType`.

In order to make it work though I needed to remove `AnExpr` from the superclasses of `DBType`. However, it was almost entirely unncessary, so this actually simplifies things as well. The only function of `AnExpr` as a superclass of `DBType` seemed to be to say "if you have an a, and a is a `DBType`, then you can do `lit a` to get an `Expr a`". However, this is kind of circular, because `DBType` by itself already gives you `encode` which is basically `lit`. It's easier to just define a columnar `lit` (which I call `monolit`) that's defined directly in terms of `encode`, and to use that in cases where you know you have a single column and going through `Serializable` doesn't quite fit.