Simplify Column type family
Now that the Field type family only needs a Necessity and a Type, we don't need to "parse" the Type to find the Necessity, we can instead just make it so that Column creates a Required field, and repurpose the name Default to be a type family that creates an Optional field.

The previous "parsing" approach was brittle sometimes, particularly in the case of polymorphism.