Add `Rel8.Schema.FieldName` and `Rel8.Table.TableFunctor`
This new context allows a data type to contain "pointers" to fields in another type. For example, in the context `FieldName (Expr Bool, Expr Bool)`, a single column can refer to either the first or second element of this tuple. A `FieldName a x` refers to a field in the type `a` of type `x`, but is otherwise an opaque pointer. The only things a user can really do is pattern match and re-arrange fields - exactly what we need for projections.

This commit also introduces `Rel8.Table.TableFunctor`, which describes the class of `Table -> Table` functors. `ListTable` and `NonEmptyTable` are instances of `TableFunctor` allowing their selected columns to be restricted or rearranged.