Add context parameter to {Either,Maybe,These}Table
This allows to get rid of all the `Tag` nonsense and to further simplify the `Nullifiable` nonsense.

Arguably we should have done it like this in the first place, as it brings a lot of nonsense that was hiding inside `Tag` out into the open.

It was nice to not have an extra redundant-seeming type parameter (because the `a` already has a `Context`), but it isn't actually redundant.

The proposed solution to the `Projection` problem will definitely need this extra `context` parameter and there are no tricks with `Tag` we can do to get around that.