Remove Interpretation and Col (#107)
This commit removes the `Interpretation` type class and - by extension - the associated `Col` newtype. We were doing this because we want to use things like `Expr` with `Type`s (e.g., `Expr Bool`), but also use `Expr`s with `Spec`s (e.g., `Col Expr ('Spec Bool)`). I realised that we don't need to jump through `Col` to do this, we can take advantage of `Expr`s polymorphic kind. This essentially means moving the constructors in `Col` directly into the associated contexts.

With this done, we no longer need to distinguish between a `HContext` and `Context`, so we drop `HContext` and change `Context` to be `Spec -> Type`. A consequence of this is that `KRel8able` changes to be `(Spec -> Type) -> Type`, which is direct conflict with `HTable` - so only one of these can be `Table`s. We choose `Rel8able` to win, and add `Cols` to treat any `HTable` as a `Table` (thus allowing any `htable f` to be a `Table f`).