Add NullTable (#173)
This is an alternative to `MaybeTable` that doesn't use a tag columns. It's less flexible (no `Functor` or `Applicative` instance) and is meaningless when used with a table that has no non-nullable columns (so nesting `NullTable` is redundant). But in situations where the underlying `Table` does have non-nullable columns, it can losslessly converted to and from `MaybeTable`.

It is useful for embedding into a base table when you don't want to store the extra tag column in your schema.

Co-authored-by: Ollie Charles <ollie@ocharles.org.uk>