Add support for prepared statements (#369)
This requires the introduction of a new `Encoder` type with three members: `binary`, which is the Hasql binary encoder, `text` which encodes a type in PostgreSQL's text format (needed for nested arrays) and `quote`, which does the thing that the function we previously called `encode` does (i.e., `a -> Opaleye.PrimExpr`).

This will break any downstream code that uses a completely custom `DBType` implementation, but anything that uses `ReadShow`, `Enum`, `Composite`, `JSONBEncoded` or `parseTypeInformation` will continue working as before (which should cover all common cases).

I dropped support for `network-ip` instead of trying to figure out how to write binary and text encoders for it. We still support `iproute`.